requirements.txt
wakeonlan
pyyaml
flask-login